Portfolio

+++ Dashboard/catlist.php
@@ -51,13 +51,13 @@
                             <div class="sb-sidenav-menu-heading">Interface</div>
                             <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#collapseLayouts" aria-expanded="false" aria-controls="collapseLayouts">
                                 <div class="sb-nav-link-icon"><i class="fas fa-columns"></i></div>
-                                Resume
+                                Education
                                 <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
                             </a>
                             <div class="collapse" id="collapseLayouts" aria-labelledby="headingOne" data-bs-parent="#sidenavAccordion">
                                 <nav class="sb-sidenav-menu-nested nav">
-                                    <a class="nav-link" href="resume.php">Resume Insert</a>
-                                    <a class="nav-link" href="resumelist.php">Resume List</a>
+                                    <a class="nav-link" href="resume.php">Educational Information Insert</a>
+                                    <a class="nav-link" href="resumelist.php">Educational Information List</a>
                                 </nav>
                             </div>
                             <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#collapsePages" aria-expanded="false" aria-controls="collapsePages">
@@ -94,13 +94,13 @@
                             </div>
                             <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#pagescollapseLayouts" aria-expanded="false" aria-controls="pagescollapseLayouts">
                                 <div class="sb-nav-link-icon"><i class="fas fa-list-alt"></i></div>
-                                Category
+                                Project Information
                                 <div class="sb-sidenav-collapse-arrow"><i class="fas fa-angle-down"></i></div>
                             </a>
                             <div class="collapse" id="pagescollapseLayouts" aria-labelledby="heading" data-bs-parent="#sidenavAccordion">
                                 <nav class="sb-sidenav-menu-nested nav">
-                                    <a class="nav-link" href="catinsert.php">Category Insert</a>
-                                    <a class="nav-link" href="catlist.php">Category List</a>
+                                    <a class="nav-link" href="catinsert.php">Project Information Insert</a>
+                                    <a class="nav-link" href="catlist.php">Project Information List</a>
                                 </nav>
                             </div>
                             <a class="nav-link collapsed" href="#" data-bs-toggle="collapse" data-bs-target="#pagescollapseLayout" aria-expanded="false" aria-controls="pagescollapseLayout">
@@ -141,7 +141,7 @@
             <div id="layoutSidenav_content">
                 <main>
                     <div class="container-fluid px-4">
-                        <h1 class="mt-4 d-flex justify-content-center">Portfolio Details</h1>
+                        <h1 class="mt-4 d-flex justify-content-center">Project Information Details</h1>
                         <div class="col-md-12">
                             <div class="input-field">
                                 <!-- Inner Join Start-->
@@ -158,7 +158,7 @@
                                 <table class="table table-striped table-bordered" style="width:100%" id="datatable">
                                     <thead>
                                         <tr>
-                                            <th>Details Id</th>
+                                            <th>Project Id</th>
                                             <th>Category Name</th>
                                             <th>Client Name</th>
                                             <th>Project Date</th>
@@ -179,25 +179,22 @@
                                                 <td><?php echo $row["projectDate"];?></td>
                                                 <td><img src="../photos/category/<?php echo $row["image"];?>" height="100" width="100" alt="<?php echo $row["image"];?>" title="<?php echo $row["image"];?>"/></td>
                                                 <td>
-                                                    <a href="catupdate.php?id=<?php echo $row["cdId"];?>">
-                                                        <button>Update</button>
-                                                    </a>
-                                                    <a href="catdelete.php?id=<?php echo $row["cId"];?>">
-                                                        <button>Delete</button>
-                                                    </a> 
+                                                    <a class="btn btn-info" href="catupdate.php?id=<?php echo $row["cdId"];?>">Update</a>
+                                                    <a class="btn btn-danger" onclick="return confirm('Are you sure to delete this item?')" href="catdelete.php?id=<?php echo $row["cId"];?>">Delete</a> 
                                                 </td>
                                             </tr>
                                         <?php }}   ?>
                                     </tbody>
-                                    <!-- <tfoot>
+                                    <tfoot>
                                         <tr>
-                                            <th>Details Id</th>
+                                            <th>Project Id</th>
                                             <th>Category Name</th>
                                             <th>Client Name</th>
                                             <th>Project Date</th>
                                             <th>Image</th>
+                                            <th>Status</th>
                                         </tr>
-                                    </tfoot> -->
+                                    </tfoot>
                                 </table>
                             </div>
                         </div>
